Add files via upload
The main objective of this work is to reduce the human death rate in road accident. The paper proposed a system to give quick assistance to the people who got the accident. The fall detection and reporting system for the vehicle can gain the attention because the system will save the life and give medical treatment on time. The system consists of Nodemcu micro-controller unit, MEMS accelerometer,crash controller, GPS device, sms via website . An Accelerometer is used to detect the acceleration. It is the main sensor used to detect the accident.The crash sensor gets active too,measure this acceleration and relay it to the control unit as usable data.Once the accident is detected GPS collect the current position values which include latitude (N or S), longitude (E or W), date and time. The location values are given to microcontroller. Controller gives this informatation. By using sms via website we can send the message to family members or EMS. The RS232 communication standard is used for the electrical signal characteristics such as voltage levels. This communication enables point to point data transfer. A high performance 16/32 bit microcontroller unit is used to process and store real time signal from the accelerometer and various sensor . All the data of these sensor and GPS data are stored in memory card for analysis of accident cause. The total system is placed inside a vehicle which is not visible to others. We can implement robust package design so that system is safe from water and dust